+++ src/test/java/com/riva/odos/services/FacilityServiceTest.java
@@ -2,25 +2,18 @@ package com.riva.odos.services;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.mockito.Spy;
 
 import static org.mockito.Mockito.doReturn;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.riva.odos.domain.FacilityDto;
 
@@ -34,18 +27,11 @@ class FacilityServiceTest {
 	@Autowired
 	FacilityService unSpyFacilityService;
 	
-	@MockBean 
-	ObjectMapper objectMapper;
-	
-	@Mock
-	JsonProcessingException jsonProcessingException;
-
 	private static final String UNKNOWN = "unknown";
 	
 	@Test
 	void shouldGetFacilities() throws Exception {
 		List<FacilityDto> facilities = unSpyFacilityService.getFacilities();
-		
 		
 		assertTrue(facilities.size() > 0);
 	}
@@ -67,26 +53,6 @@ class FacilityServiceTest {
 		assertEquals(expectedfacilityInfo.getZipCode(), returnedFacilityList.get(0).getZipCode());
 	}
 	
-	@Test
-	void testSearchFacilitiesExceptionThrown() throws Exception {
-		Mockito.doAnswer(invocation -> {
-            throw jsonProcessingException;
-        }).when(objectMapper).readValue(Mockito.anyString(), Mockito.any(TypeReference.class));
-		unSpyFacilityService.getFacilities();
-	}
-	
-//	@Test
-//	void testRetrieveJsonExceptionThrown() throws Exception {
-//		Mockito.doAnswer(invocation -> {
-//            throw new IOException();
-//        }).when(objectMapper).readValue("", new TypeReference<List<FacilityDto>>() {});
-//		
-//	}
-	
-	@Test
-	void testRetrieveJsonException() {
-		
-	}
 	
 	@Test
 	void testSearchFacilitiesNotFound() throws Exception{
